Move tests from lib.rs to tests.rs
